Move message reading internals into AbstractClientConnection
`AbstractConnection` provided utility methods for reading messages from a stream, decompressing them, and dispatching to
handlers. This has all been moved into `AbstractClientConnection` instead.
